+++ api/controllers/auth.controller.js
@@ -177,6 +177,18 @@ exports.signin = (req, res) => {
       });
 };
 
+exports.googleSignIn = function(user) {
+  var token = jwt.sign({ id: user.id }, config.secret, {
+    expiresIn: 86400 // 24 hours
+  });
+
+  let obj = {
+    id: user._id,
+    username: user.username,
+    accessToken: token
+  }
+}
+
 exports.changePasswordUsingMail = (req, res) => {
   var updatedUser = {
     password: bcrypt.hashSync(req.body.password, 8)